refactor(demux): :zap: move code from index file to seperated function
